+++ astropy/time/erfa_time.pyx
@@ -47,6 +47,7 @@ cdef extern from "erfa.h":
     # Geodetic
     int eraAf2a(char s, int ideg, int iamin, double asec, double *rad)
     int eraGd2gc(int n, double elong, double phi, double height, double xyz[3])
+    int eraGc2gd(int n, double xyz[3], double *elong, double *phi, double *height )
 
     # Sidereal time
     double eraGmst06(double uta, double utb, double tta, double ttb)
@@ -1124,7 +1125,9 @@ def utc_ut1(
 def d_tdb_tt(np.ndarray[double, ndim=1] in1,
              np.ndarray[double, ndim=1] in2,
              np.ndarray[double, ndim=1] ut,
-             elong, u, v):
+             np.ndarray[double, ndim=1] elong,
+	     np.ndarray[double, ndim=1] u,
+	     np.ndarray[double, ndim=1] v):
     """
     compute DTR = TDB-TT
     double eraDtdb(double date1, double date2, double ut,
@@ -1228,15 +1231,15 @@ def d_tdb_tt(np.ndarray[double, ndim=1] in1,
     **     TCB and hence between TT and TDB.
     """
     assert in1.shape[0] == in2.shape[0] == ut.shape[0]
+    assert elong.shape[0] == u.shape[0] == v.shape[0]
+    assert elong.shape[0] == 1 or elong.shape[0] == in1.shape[0]
     cdef unsigned n = in1.shape[0]
-    cdef unsigned int i
+    cdef unsigned int i, j
     cdef np.ndarray[double, ndim=1] out = np.empty(n, dtype=np.double)
-    cdef double c_elong = elong
-    cdef double c_u = u
-    cdef double c_v = v
 
     for i in range(n):
-        out[i] = eraDtdb(in1[i], in2[i], ut[i], c_elong, c_u, c_v)
+        j = min(i, elong.shape[0]-1)
+        out[i] = eraDtdb(in1[i], in2[i], ut[i], elong[j], u[j], v[j])
     return out
 
 
@@ -1325,16 +1328,91 @@ def era_gd2gc(n, elong, phi, height):
     **
     **  4) The inverse transformation is performed in the function eraGc2gd.
     """
+    assert elong.shape[0] == phi.shape[0] == height.shape[0]
+    cdef unsigned int i
+    cdef unsigned int nitems = elong.shape[0]
     cdef np.ndarray[double, ndim=1] xyz = np.empty(3, dtype=np.double)
+    cdef np.ndarray[double, ndim=2] out = np.empty((nitems, 3), dtype=np.double)
 
     errs = {-1: 'illegal identifier',
              -2: 'illegal case'}
 
-    ret = eraGd2gc(n, elong, phi, height, &xyz[0])
-    check_return(ret, 'eraGd2gc', errors=errs)
+    for i in range(nitems):
+        ret = eraGd2gc(n, elong[i], phi[i], height[i], &xyz[0])
+        check_return(ret, 'eraGd2gc', errors=errs)
+        out[i] = xyz
 
-    return xyz
+    return out
 
+@cython.wraparound(False)
+@cython.boundscheck(False)
+def era_gc2gd(n, xyz):
+    """
+    Wrap
+    int eraGc2gd(int n, double xyz[3], double *elong, double *phi, double *height )
+
+    **  Given:
+    **     n       int        ellipsoid identifier (Note 1)
+    **     xyz     double[3]  geocentric vector (Note 2)
+    **
+    **  Returned:
+    **     elong   double     longitude (radians, east +ve)
+    **     phi     double     latitude (geodetic, radians, Note 3)
+    **     height  double     height above ellipsoid (geodetic, Notes 2,3)
+    **
+    **  Returned (function value):
+    **            int         status:  0 = OK
+    **                                -1 = illegal identifier (Note 3)
+    **                                -2 = internal error (Note 3)
+    **
+    **  Notes:
+    **
+    **  1) The identifier n is a number that specifies the choice of
+    **     reference ellipsoid.  The following are supported:
+    **
+    **        n    ellipsoid
+    **
+    **        1     ERFA_WGS84
+    **        2     ERFA_GRS80
+    **        3     ERFA_WGS72
+    **
+    **     The n value has no significance outside the ERFA software.  For
+    **     convenience, symbols ERFA_WGS84 etc. are defined in erfam.h.
+    **
+    **  2) The geocentric vector (xyz, given) and height (height, returned)
+    **     are in meters.
+    **
+    **  3) An error status -1 means that the identifier n is illegal.  An
+    **     error status -2 is theoretically impossible.  In all error cases,
+    **     phi and height are both set to -1e9.
+    **
+    **  4) The inverse transformation is performed in the function eraGd2gc.
+    **
+    **  Called:
+    **     eraEform     Earth reference ellipsoids
+    **     eraGc2gde    geocentric to geodetic transformation, general
+    **
+    **  Copyright (C) 2013, NumFOCUS Foundation.
+    **  Derived, with permission, from the SOFA library.  See notes at end of file.
+    """
+    assert xyz.shape[1] == 3
+    cdef unsigned int i
+    cdef unsigned int nitems = xyz.shape[0]
+    cdef np.ndarray[double, ndim=1] elong = np.empty(nitems, dtype=np.double)
+    cdef np.ndarray[double, ndim=1] phi = np.empty(nitems, dtype=np.double)
+    cdef np.ndarray[double, ndim=1] height = np.empty(nitems, dtype=np.double)
+    cdef double xyz_item[3]
+
+    errs = {-1: 'illegal identifier',
+            -2: 'illegal case'}
+    for i in range(nitems):
+        # ensure xyz are in a contiguous array
+        for j in range(3):
+            xyz_item[j] = xyz[i, j]
+        ret = eraGc2gd(n, xyz_item, &elong[i], &phi[i], &height[i])
+        check_return(ret, 'eraGd2gc', errors=errs)
+
+    return elong, phi, height
 
 @cython.wraparound(False)
 @cython.boundscheck(False)