+++ CHANGES.rst
@@ -211,6 +211,10 @@ API Changes
 
   - ``TimeDelta`` objects can have scales other than TAI [#1932].
 
+  - Location information should now be passed on via an ``EarthLocation``
+    instance or anything that initialises it, e.g., a tuple containing
+    either geocentric or geodetic coordinates. [#1928]
+
 - ``astropy.units``
 
   - ``Quantity`` now converts input to float by default, as this is physically